chore: move SettledUpdate type to the types package (#1689)
